Play Presets from rundown JSON
I added an API to SPX-GC to get the current Rundown JSON.                   
With this code there will be presets that are generated from the Rundown    
This will make it easier to use for users because they don't need to manually read the JSON file to get the Item ID's.

I added some Tests to make sure I don't break existing Presets.

In this Version, it will only get the Rundown on first connect/reconnect.
I hope I can change that to a polling approach later, or even use the socket.io used in SPX-CG internally.